Conflict with Box2D resolved
*had using namespace glm; at the top of files to save from explicitly
naming the namespace for everything. This caused a ambiguity error when
Box2D was included because both glm and Box2D have Uint as a type def
for unsigned int. When the namespace was named ahead of time, the
compiler wasn't sure whether to use the glm Uint in the Box2D files. The
using namespace glm; has been removed now and all is well.